requirements.txt
asgiref==3.2.10
Django==3.0.7
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
django-import-export==2.3.0  